+++ src/Commands/SetUpAndUseTailwindCli.cs
@@ -54,7 +54,7 @@ internal sealed class SetUpAndUseTailwindCli : BaseCommand<SetUpAndUseTailwindCl
             settings.ConfigurationFiles.Add(new() { Path = configFile! });
             settings.BuildFiles.Add(new() { Input = configFile! });
             settings.UseCli = true;
-            await SettingsProvider.OverrideSettingsAsync(settings);
+            await SettingsProvider.OverrideSettingsAsync(settings, directory);
 
             var file = await PhysicalFile.FromFileAsync(SolutionExplorerSelection.CurrentSelectedItemFullPath);